Regression Discontinuity Design Permutation Test

A permutation test for continuity of covariates in Sharp Regression Discontinuity Design as described in Canay and Kamat (2017).

RDperm(W, z, data, n.perm = 499, q_type = 10, cutoff = 0,
  test.statistic = "CvM")

Arguments

W
Character. Vector of covariates names. The procedure will test the null hypothesis of continuity of the distribution of each element in W at the cutoff.

z
Character. Running variable name. This is the scalar random variable that defines, along with the cutoff, the treatment assignment rule in the sharp regression discontinuity design.

data
Data.frame.

n.perm
Numeric. Number of permutations needed for the stochastic approximation of the p-values. See remark 3.2 in Canay and Kamat (2017). The default is B=499.

q_type
A fixed and small (relative to the sample size) natural number that will define the \(q\) closest values of the order statistic of \(Z\) to the right and to the left of the cutoff. The default, 'rot', value is given by the feasible rule of thumb in footnote 4 of Canay and Kamat (2017), section 3.1. If 'arot', it calls for the Rule of Thumb described in equation (15) of Canay and Kamat (2017), section 3.1. The default option grows at a slower rate than the optional rule of thumb, but adds a larger constant.

cutoff
Numeric. The scalar defining the threshold of the running variable.

test.statistic
Character. A rank test statistic satisfying rank invariance. The default is a Cramer-von Mises test statistic.

Value

The functions summary and plot are used to obtain and print a summary and plot of the estimated regression discontinuity. The object of class RDperm is a list containing the following components:

results

Matrix. Test Statistic, P-values and Q

test.statistic

Test Statistic

q_type

Type of Q used in the calculations, can be either, "Defined by User", the "Rule of Thumb" or the "Alternative Rule of Thumb".

n_perm

number of permutations

rv

Character. Running variable name

Z

Vector. Running Variable

cutoff

cutoff

data

data set

S

Matrix. Pooled sample of induced order statistics

S_perm

List. Permutations of the induced order statistic.

References

Canay, I and Kamat V, (2017) Approximate Permutation Tests and Induced Order Statistics in the Regression Discontinuity Design. http://faculty.wcas.northwestern.edu/~iac879/wp/RDDPermutations.pdf

Examples

permtest<-RDperm(W=c("demshareprev"),z="difdemshare",data=lee2008)
summary(permtest)

#> 
#> **********************************************************
#> **       RD Distribution Test using permutations        **
#> **********************************************************
#> Running Variable: difdemshare
#> Cutoff: 0
#> q: Defined by User
#> Test Statistic: CvM
#> Number of Permutations: 499
#> Number of Obs: 6558
#> 
#> **********************************************************
#> H0: 'Continuity of the baseline covariates at the cutoff' 
#> **********************************************************
#> 
#> Estimates:
#>               T(Sn)  Pr(>|z|)  q    
#> demshareprev  0.09   0.05      10   
#> ---
#> Signif. codes:   0.01 '***' 0.05 '**' 0.1 '*' 
#> 
# NOT RUN {
permtest<-RDperm(W=c("demshareprev","demwinprev"),z="difdemshare",data=lee2008)
summary(permtest)
# }